REFACTOR: move one level up FIX: test case super class
