+++ tests/whist/server/base_user_test_case.py
@@ -1,9 +1,11 @@
+import unittest
+
 from whist.server.database import db
 from whist.server.database.user import UserInDb
 from whist.server.services.user_db_service import UserDatabaseService
 
 
-class UserBaseTestCase:
+class UserBaseTestCase(unittest.TestCase):
     def setUp(self):
         self.user_database_service = UserDatabaseService()
         self.user: UserInDb = UserInDb(username='test', hashed_password='abc')
